Bugfix: dynamically set ChoiceField choices in DescriptionForm
Resolved a `ProgrammingError` encountered during initial database
creation and migration by moving the setting of `choices` for the
`name` field in `DescriptionForm` from class definition to the
`__init__` method. This change ensures that database queries for
populating `choices` are made at runtime, avoiding attempts to access
the `describe_description` table before it exists in the database.